+++ site/_data/stages.yaml
@@ -1,0 +1,75 @@
+---
+
+-
+    name: introduction
+
+    instructions: |
+        <h1> Welcome to Prime Image </h1>
+        
+        <p> Click <button class="change-stage" data-stage-change="1">Next</button> to begin. </p>
+    interface: ''
+    display: |
+        <span class="image-number">1001010777777777777770777777777777777777770777777777777777777777
+        1010010777777777777702077777777777777777702077777777777777777777
+        1001000777777777777770777777777777777777770777777777777777777777
+        7777777777777777777705077777777777777777705077777777777777777777
+        7777777777777777777055107777777777777777015507777777777777777777
+        7777777777777777777055507777777777777777055507777777777777777777
+        7777777777777777777000037777777007777777300007777777777777777777
+        7777777777777777770433307777770880777777033340777777777777777777
+        7777777778777777777000007777771001777777000007777777777777777777
+        7778777077777777777099907777710550177777099907777777777707777777
+        7777777017777777777099907777105555017777099907777777777107777777
+        7777777047777777777099907717053003507177100007777777777407777777
+        7777770107777777787000800500511111150050080007777777777010777777
+        7777770107777777777109900051100000011500099907777777777010777777
+        7777777247777777777000001115051111605111001007777777777327777777
+        7777777007777787777099301500115005110051039907777777777007777777
+        7777777077877777771099300011110880111100039901777787777707777777
+        7777777047777777700000001111115005111112000000078777777407778777
+        7000700560110011055059901111111111111111099505501100110650070007
+        0656065555005500555009001111116005111111009005550055005555606560
+        5555555555555555555009001115088988805111009005555555555555555555
+        0000005500000000000090901119088888809111090900000000000055000000
+        1111105501111111111090901110888888880111090901111111111055011111
+        9511105501115995111090901118008008008111090901115995111055011159
+        0851105501168008511030301118888888888111030301159008512065011580
+        0801100001108008011053501118888888889111053501108008011000011080
+        0801105501108008010900001118888888888111000090108018011055011080
+        0802105501108008011055501110000330000111056501108008011055011080
+        0801105501108008011059503400000000099933059501108008011055011080
+        3331105501133333311093305555655555555555033901133333312055011333
+        0000005500000000000095909999999999999999095900000000000055000000
+        1111505505111111111094901111112111111111094901111111115055051111
+        1111105501111111111092901119000100009111092901111111111055011112
+        1111100001111112111095901210888988880111095901111111111000011111
+        8511105501115891111095901110800880080111195901111985111055011158
+        0051106501150008111092901110088888800111092901118000511055011500
+        0831105501138080110055901110188888800111095500110808311055011380
+        0831105501138080110555550110088888800110555550110808311056011390
+        0031105611130000110555550110088888800110555550110000311055011300
+        1111055550111111110555550110088888800110555550111111110555501112
+        1111055550111111110555550110088888800110555550111111110550000000
+        0000055550000000000555550000088888800000555550000000000550120397</span>
+-
+    name: image-load
+    instructions: |
+        <h1> Select Image </h1>
+        
+        <p> First you must select an image to turn into a prime number. </p>
+        
+        <p> Use the box below to open an image. </p>
+    display: |
+        <canvas></canvas>
+    
+    interface: |
+      <form class="image-pick">
+        <svg class="icon" xmlns="http://www.w3.org/2000/svg" width="50" height="43" viewBox="0 0 50 43">
+          <path d="M48.4 26.5c-.9 0-1.7.7-1.7 1.7v11.6h-43.3v-11.6c0-.9-.7-1.7-1.7-1.7s-1.7.7-1.7 1.7v13.2c0 .9.7 1.7 1.7 1.7h46.7c.9 0 1.7-.7 1.7-1.7v-13.2c0-1-.7-1.7-1.7-1.7zm-24.5 6.1c.3.3.8.5 1.2.5.4 0 .9-.2 1.2-.5l10-11.6c.7-.7.7-1.7 0-2.4s-1.7-.7-2.4 0l-7.1 8.3v-25.3c0-.9-.7-1.7-1.7-1.7s-1.7.7-1.7 1.7v25.3l-7.1-8.3c-.7-.7-1.7-.7-2.4 0s-.7 1.7 0 2.4l10 11.6z">
+          </path>
+        </svg>
+        <input type="file" name="files[]" id="file" />
+        <label for="file"><strong>Choose a file</strong> or drag it here.</label>
+      </form>
+---
+
