+++ site/_includes/stage-content.html
@@ -1,0 +1,6 @@
+
+{% for md in include.stages %}
+<{{ include.element }}>
+  {{ md }}
+</{{ include.element }}>
+{% endfor %}
