+++ site/_sass/_display.scss
@@ -1,0 +1,7 @@
+.image-number {
+  font-size: 0.35em;
+  line-height: 0.8em;
+  font-weight: bold;
+  text-align: center;
+  white-space: pre;
+}
