+++ site/_sass/_instructions.scss
@@ -1,0 +1,10 @@
+h1 {
+  color: white;
+  font-size: 1.3em;
+  margin-top: 0;
+}
+
+p {
+  background: white;
+  padding: 0.6em 0.3em;
+}
