+++ site/index.html
@@ -1,12 +1,9 @@
 ---
 layout: default
 title: Prime Image
-js-module:
-  - prime-image.js
-stages:
-  - intro
-  - image-load
 ---
+
+{% assign stages = site.data.stages %}
 
 <header>
   <h1> {{ page.title }} </h1>
@@ -15,27 +12,12 @@ stages:
 <main>
   <div class="interaction-panel" id="interaction">
     <section class="interaction-instructions stage-selecting show-only-selected">
-      {% for stage in page.stages %}
-      <span>
-        {%capture include_instructions %}
-          {% include_relative stages/{{stage}}/instructions.md %}
-        {% endcapture %}
-        {{ include_instructions | markdownify }}
-      </span>
-      {% endfor %}
-    
+      {% assign instructions = stages | map: "instructions" %}
+      {% include stage-content.html stages=instructions element="span" %}
     </section>
     <div class="interaction-interface stage-selecting show-only-selected">
-      <span></span>
-      <form class="image-pick">
-        <svg class="icon" xmlns="http://www.w3.org/2000/svg" width="50" height="43" viewBox="0 0 50 43">
-          <path d="M48.4 26.5c-.9 0-1.7.7-1.7 1.7v11.6h-43.3v-11.6c0-.9-.7-1.7-1.7-1.7s-1.7.7-1.7 1.7v13.2c0 .9.7 1.7 1.7 1.7h46.7c.9 0 1.7-.7 1.7-1.7v-13.2c0-1-.7-1.7-1.7-1.7zm-24.5 6.1c.3.3.8.5 1.2.5.4 0 .9-.2 1.2-.5l10-11.6c.7-.7.7-1.7 0-2.4s-1.7-.7-2.4 0l-7.1 8.3v-25.3c0-.9-.7-1.7-1.7-1.7s-1.7.7-1.7 1.7v25.3l-7.1-8.3c-.7-.7-1.7-.7-2.4 0s-.7 1.7 0 2.4l10 11.6z">
-          </path>
-        </svg>
-
-        <input type="file" name="files[]" id="file" />
-        <label for="file"><strong>Choose a file</strong> or drag it here.</label>
-      </form>
+      {% assign interface = stages | map: "interface" %}
+      {% include stage-content.html stages=interface element="span" %}
     </div>
     <div class="interaction-control">
       <button disabled class="fake-button change-stage" data-stage-change="-1">Back</button>
@@ -44,13 +26,8 @@ stages:
   </div>
   <div>
     <div class="display-panel stage-selecting change-stage" data-stage-change="1">
-    
-      {% for stage in page.stages %}
-      <div>
-        {% include_relative stages/{{stage}}/display.html %}
-      </div>
-      {% endfor %}
-    
+      {% assign display = stages | map: "display" %}
+      {% include stage-content.html stages=display element="div" %}
     </div>
   </div>
 </main>
