--- site/stages/image-load/display.html
@@ -1,1 +1,0 @@
-<canvas></canvas>
