--- site/stages/image-load/instructions.md
@@ -1,8 +1,0 @@
----
----
-
-# Select Image
-
-First you must select an image to turn into a prime number.
-
-Use the box below to open an image.
