--- site/stages/intro/instructions.md
@@ -1,6 +1,0 @@
----
----
-
-# Welcome to Prime Image
-
-Click <button class="change-stage" data-stage-change="1">Next</button> to begin.
