--- site/stages/intro/style
@@ -1,5 +1,0 @@
-
-font-size: 0.5em;
-line-height: 1em;
-font-weight: bold;
-text-align: center;
